modCliente terminado

+++ src/GrouponDesktop/AbmCliente/AltaCliente.cs
@@ -150,7 +150,7 @@ namespace GrouponDesktop.AbmCliente
             }
             else
             {
-                if (!validarTelUnico(Convert.ToInt32(this.masktxtTel.Text)))
+                if (!validarTelUnico(Convert.ToInt64(this.masktxtTel.Text)))
                 {
                     lblTel.ForeColor = System.Drawing.Color.Red;
                     strError += "- Ya hay registrado un cliente con este telefono\n";
@@ -182,7 +182,7 @@ namespace GrouponDesktop.AbmCliente
         }
 
 
-        private bool validarTelUnico(Int32 tel)
+        private bool validarTelUnico(Int64 tel)
         {
             bool telUnico = false;
             SqlConnection dbcon = new SqlConnection(GrouponDesktop.Properties.Settings.Default["conStr"].ToString());
@@ -190,7 +190,7 @@ namespace GrouponDesktop.AbmCliente
 											from LOSGROSOS_RELOADED.Clientes
 											where tel=@tel", dbcon);
 
-            cmd.Parameters.Add("@tel", SqlDbType.Int, 18);
+            cmd.Parameters.Add("@tel", SqlDbType.BigInt, 18);
             cmd.Parameters["@tel"].Value = tel;
             DataTable dt = new DataTable();
             SqlDataAdapter da = new SqlDataAdapter(cmd);
@@ -235,7 +235,7 @@ namespace GrouponDesktop.AbmCliente
             cmd.Parameters.Add("@apellido", SqlDbType.VarChar, 255);
             cmd.Parameters.Add("@dni", SqlDbType.Int, 18);
             cmd.Parameters.Add("@direccion", SqlDbType.VarChar, 255);
-            cmd.Parameters.Add("@tel", SqlDbType.Int, 18);
+            cmd.Parameters.Add("@tel", SqlDbType.BigInt, 18);
             cmd.Parameters.Add("@mail", SqlDbType.VarChar, 255);
             cmd.Parameters.Add("@fechaNac", SqlDbType.DateTime);
             cmd.Parameters.Add("@idCiudad", SqlDbType.Int, 18);
@@ -247,7 +247,7 @@ namespace GrouponDesktop.AbmCliente
             cmd.Parameters["@apellido"].Value = this.txtApellido.Text;
             cmd.Parameters["@dni"].Value = Convert.ToInt32(this.txtDni.Text);
             cmd.Parameters["@direccion"].Value = this.txtDireccion.Text;
-            cmd.Parameters["@tel"].Value = Convert.ToInt32(this.masktxtTel.Text);
+            cmd.Parameters["@tel"].Value = Convert.ToInt64(this.masktxtTel.Text);
             cmd.Parameters["@mail"].Value = this.txtMail.Text;
             cmd.Parameters["@idCiudad"].Value = idCiudad;
             cmd.Parameters["@saldo"].Value = Support.saldoBienvenida;
